Refactor: Split RealtimeDataService into smaller files
Split RealtimeDataService.ts into smaller, more manageable files while ensuring no functionality is broken. Removed unused imports and files.
